% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{accept_prob_est_gauss}
\alias{accept_prob_est_gauss}
\title{Acceptance Probability Estimation with Gaussian Envelope}
\usage{
accept_prob_est_gauss(mu, sd, log_target_density)
}
\arguments{
\item{mu}{Mean of the Gaussian envelope.}

\item{sd}{Standard deviation of the Gaussian envelope.}

\item{log_target_density}{Function computing the log target density.}
}
\value{
A numeric estimate of the acceptance probability \code{alpha}.
}
\description{
Estimates the acceptance probability \code{alpha} using a Gaussian envelope.
}
