% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{adaptive_one_time_placement}
\alias{adaptive_one_time_placement}
\title{Perform Adaptive One-Time Placement}
\usage{
adaptive_one_time_placement(
  initial_envelope,
  log_target_density,
  d_log_target_density,
  initial_samples = 1000,
  num_knots = 5,
  ...
)
}
\arguments{
\item{initial_envelope}{Initial envelope object.}

\item{log_target_density}{Function for computing the log target density.}

\item{d_log_target_density}{Function for computing the derivative of the log target density.}

\item{initial_samples}{Number of initial samples.}

\item{num_knots}{Number of knots to place adaptively.}
}
\value{
A list containing the modified envelope, knot positions, and initial samples.
}
\description{
Adjusts knot placement in the envelope function based on initial samples from rejection sampling.
}
