% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{add_density_to_plot}
\alias{add_density_to_plot}
\title{Add Density Line to Plot}
\usage{
add_density_to_plot(plot, density_object, color = "red")
}
\arguments{
\item{plot}{A ggplot object to which the density line will be added.}

\item{density_object}{A list with two components: \code{x} (numeric vector for x-coordinates) and \code{y} (numeric vector for y-coordinates).}

\item{color}{A character string specifying the color of the density line. Default is \code{"red"}.}
}
\value{
A ggplot object with the added density line.
}
\description{
This helper function adds a density line to an existing ggplot object based on a specified density.
}
