% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{autoplot}
\alias{autoplot}
\title{Plot Density and Histogram for Custom Density Object}
\usage{
autoplot(
  density_object,
  histogram = TRUE,
  true_density_function = NULL,
  color = "black"
)
}
\arguments{
\item{density_object}{A custom density object with components \code{data} (numeric vector of sample points), \code{x} (x-coordinates for density), and \code{y} (y-coordinates for density).}

\item{histogram}{A logical value indicating if a histogram should be added to the plot. Default is \code{TRUE}.}

\item{true_density_function}{An optional function to compute the true density values for comparison. Default is \code{NULL}.}

\item{color}{A character string specifying the color of the density line for \code{density_object}. Default is \code{"black"}.}
}
\value{
A ggplot object showing the density and optional histogram with a true density line.
}
\description{
Generates a ggplot visualization of a custom density object, optionally adding a histogram and/or a true density line.
}
