% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{bandwidth_silvermans_rule}
\alias{bandwidth_silvermans_rule}
\title{Title: Silverman's Rule of Thumb for Bandwidth}
\usage{
bandwidth_silvermans_rule(vec_obs)
}
\arguments{
\item{vec_obs}{A numeric vector of observed data points.}
}
\value{
A numeric value representing the estimated bandwidth.
}
\description{
Description: Computes the bandwidth for kernel density estimation using Silverman's rule of thumb.
}
