% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binning_cpp}
\alias{binning_cpp}
\title{Data Binning Function (C++)}
\usage{
binning_cpp(vec_obs, lower, upper, number_of_bins)
}
\arguments{
\item{vec_obs}{A numeric vector of observed data points.}

\item{lower}{The lower bound of the binning range.}

\item{upper}{The upper bound of the binning range.}

\item{number_of_bins}{An integer specifying the number of bins.}
}
\value{
A numeric vector with counts of observations in each bin.
}
\description{
Bins observed data points within a specified range, dividing the range into equal-width bins.
}
