% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{calculate_oracle_bw_epanechnikov}
\alias{calculate_oracle_bw_epanechnikov}
\title{Title: Oracle Bandwidth for Epanechnikov Kernel}
\usage{
calculate_oracle_bw_epanechnikov(vec_obs)
}
\arguments{
\item{vec_obs}{A numeric vector of observed data points.}
}
\value{
A numeric value representing the oracle bandwidth for the Epanechnikov kernel.
}
\description{
Description: Computes an optimal (oracle) bandwidth for kernel density estimation with the Epanechnikov kernel.
}
