% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{controller.cpp}
\alias{controller.cpp}
\title{Control Sampling Process Using C++ Vectorized Method}
\usage{
\method{controller}{cpp}(setup, sample_size, scaling_factor = 1, min_proposals = 1)
}
\arguments{
\item{setup}{An object of class \code{gaussian_envelope_class}, containing setup information and methods for sampling.}

\item{sample_size}{An integer specifying the total number of samples to generate.}

\item{scaling_factor}{A numeric value to scale the sampling process. Default is \code{1}.}

\item{min_proposals}{An integer specifying the minimum number of proposals required. Default is \code{1}.}
}
\value{
The output of \code{controller.vectorize}, which may be a vector, list, or object containing the generated samples.
}
\description{
This function serves as a controller to manage the sampling process using a C++-based vectorized method.
It verifies that the input \code{setup} is of class \code{gaussian_envelope_class} and then calls \code{controller.vectorize}
to execute the sampling procedure based on the specified parameters.
}
\seealso{
\code{\link{controller.vectorize}}
}
