% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{controller.vectorize}
\alias{controller.vectorize}
\title{controller vectorize}
\usage{
\method{controller}{vectorize}(setup, sample_size, scaling_factor = 1, min_proposals = 1)
}
\description{
controller vectorize
}
