% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{create_log_envelope_density}
\alias{create_log_envelope_density}
\title{Compute Log Envelope Density}
\usage{
create_log_envelope_density(x, z, a, b)
}
\arguments{
\item{x}{A numeric vector of values for which to compute the log envelope density.}

\item{z}{A numeric vector of interval boundaries corresponding to each element in \code{a} and \code{b}.}

\item{a}{A numeric vector of coefficients for linear transformations within each interval.}

\item{b}{A numeric vector of intercepts for linear transformations within each interval.}
}
\value{
A numeric vector representing the computed log envelope density for each value in \code{x}.
}
\description{
This function computes the log envelope density for each value in \code{x}, given a set of intervals defined by \code{z}.
The density is calculated using linear transformations specified by coefficients \code{a} and intercepts \code{b}.
}
\seealso{
\code{\link{create_sampler_function}}
}
