% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{create_sampler_function}
\alias{create_sampler_function}
\title{Generate Samples Using Custom Sampler}
\usage{
create_sampler_function(n, c, Q, a, b, z)
}
\arguments{
\item{n}{An integer specifying the number of samples to generate.}

\item{c}{A numeric constant used in scaling the cumulative probability.}

\item{Q}{A numeric vector of cumulative probabilities defining interval boundaries.}

\item{a}{A numeric vector of coefficients used in the transformation of sampled values.}

\item{b}{A numeric vector of intercepts used in the transformation of sampled values.}

\item{z}{A numeric vector defining interval start points, corresponding to each cumulative probability in \code{Q}.}
}
\value{
A numeric vector of sampled values generated according to the specified parameters.
}
\description{
This function generates samples based on a custom piecewise distribution defined by parameters \code{c}, \code{Q}, \code{a}, \code{b}, and \code{z}.
It calculates a cumulative probability and finds the interval index for each sample, then computes the sampled values
based on the specified parameters.
}
\seealso{
\code{\link{create_log_envelope_density}}
}
