% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_sampler_function_cpp}
\alias{create_sampler_function_cpp}
\title{Sampler Function with Acceptance-Rejection (C++)}
\usage{
create_sampler_function_cpp(n, c, Q, a, b, z)
}
\arguments{
\item{n}{Integer specifying the number of samples.}

\item{c}{Constant multiplier for scaling.}

\item{Q}{A cumulative probability distribution vector.}

\item{a}{A vector of coefficients.}

\item{b}{A vector of coefficients.}

\item{z}{A vector of control parameters.}
}
\value{
A numeric vector of sampled values.
}
\description{
Samples values according to a custom acceptance-rejection method.
}
