% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_log_target_density_cpp}
\alias{d_log_target_density_cpp}
\title{Derivative of Log Target Density for Vector of Observations (C++)}
\usage{
d_log_target_density_cpp(y)
}
\arguments{
\item{y}{A numeric vector of values for which to calculate the derivative of the log target density.}
}
\value{
A numeric vector of derivative values of the log target density.
}
\description{
Calculates the derivative of the log target density for each value in a numeric vector.
}
