% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{d_log_target_density_vectorize}
\alias{d_log_target_density_vectorize}
\title{Derivative of Log Target Density Function}
\usage{
d_log_target_density_vectorize(y)
}
\arguments{
\item{y}{A numeric input vector for derivative evaluation.}
}
\value{
The derivative of the log target density for each \code{y}.
}
\description{
Computes the derivative of the log target density function for a Poisson model.
}
