% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{epanechnikov_kernel}
\alias{epanechnikov_kernel}
\title{Title: Epanechnikov Kernel Function}
\usage{
epanechnikov_kernel(x)
}
\arguments{
\item{x}{A numeric value for which the kernel value is to be calculated.}
}
\value{
A numeric value representing the kernel density, which will be zero for absolute values of \code{x} greater than 1.
}
\description{
Description: Calculates the value of the Epanechnikov kernel for a given input, \code{x}.
The Epanechnikov kernel is a common kernel function used in density estimation.
}
