% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{epanechnikov_kernel_cpp}
\alias{epanechnikov_kernel_cpp}
\title{Epanechnikov Kernel Function (C++)}
\usage{
epanechnikov_kernel_cpp(x)
}
\arguments{
\item{x}{A numeric vector of values for which to calculate the Epanechnikov kernel.}
}
\value{
A numeric vector of kernel values, with zeros for inputs outside the range -1, 1.
}
\description{
Calculates the Epanechnikov kernel for a given vector of input values.
}
