% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{gaussian_kernel}
\alias{gaussian_kernel}
\title{Title: Gaussian Kernel Function}
\usage{
gaussian_kernel(x)
}
\arguments{
\item{x}{A numeric value for which the Gaussian kernel density is to be calculated.}
}
\value{
A numeric value representing the Gaussian kernel density at the input \code{x}.
}
\description{
Description: Calculates the value of the Gaussian kernel for a given input, \code{x}.
The Gaussian kernel is widely used in kernel density estimation and smoothing.
}
