% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_kernel_cpp}
\alias{gaussian_kernel_cpp}
\title{Gaussian Kernel Function (C++)}
\usage{
gaussian_kernel_cpp(x)
}
\arguments{
\item{x}{A numeric vector of values for which to calculate the Gaussian kernel.}
}
\value{
A numeric vector of Gaussian kernel values for each element in \code{x}.
}
\description{
Calculates the Gaussian (Normal) kernel for a given vector of input values.
}
