% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{get_accepts.naive}
\alias{get_accepts.naive}
\title{Compute Acceptances Using Naive Method}
\usage{
\method{get_accepts}{naive}(setup, proposals)
}
\arguments{
\item{setup}{A list containing the functions \code{setup$log_target_density} and \code{setup$log_envelope_density}.
These functions compute log densities for the target and envelope distributions, respectively.}

\item{proposals}{A numeric vector of proposals for which acceptances are to be evaluated.}
}
\value{
A logical vector indicating whether each proposal is accepted (\code{TRUE}) or rejected (\code{FALSE}).
}
\description{
This function evaluates acceptances for a set of proposals using a naive acceptance rule.
The acceptance probability is determined by comparing random uniform values to the ratio of
the target density to an envelope density, both defined in \code{setup}.
}
