% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_accepts_gaussian_cpp}
\alias{get_accepts_gaussian_cpp}
\title{Gaussian Rejection Sampler Accept/Reject Decisions (C++)}
\usage{
get_accepts_gaussian_cpp(mu, sd, log_alpha_prime, proposals, u_samples)
}
\arguments{
\item{mu}{Mean of the Gaussian distribution.}

\item{sd}{Standard deviation of the Gaussian distribution.}

\item{log_alpha_prime}{Log scaling parameter for acceptance region.}

\item{proposals}{A numeric vector of proposed sample values.}

\item{u_samples}{A numeric vector of random uniform values for rejection sampling.}
}
\value{
A logical vector indicating accepted (\code{TRUE}) or rejected (\code{FALSE}) samples.
}
\description{
Determines acceptance of samples in a Gaussian rejection sampling algorithm.
}
