% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{get_cv_bw}
\alias{get_cv_bw}
\title{Title: Cross-Validation Bandwidth Selection}
\usage{
get_cv_bw(
  vec_obs,
  number_of_partitions = 10,
  kernel_evaluator = kernel_cpp_epanechnikov
)
}
\arguments{
\item{vec_obs}{A numeric vector of observed data points.}

\item{number_of_partitions}{The number of partitions for cross-validation (default is 10).}

\item{kernel_evaluator}{A kernel function, typically an optimized C++ version for efficiency.}
}
\value{
A numeric value representing the CV-selected bandwidth.
}
\description{
Description: Calculates the optimal bandwidth for kernel density estimation using cross-validation.
}
