% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{get_em_setup}
\alias{get_em_setup}
\title{Set Up EM Optimization}
\usage{
get_em_setup(x)
}
\arguments{
\item{x}{A numeric vector of observed data.}
}
\value{
A list structure for EM containing \code{e_step}, \code{m_step}, and \code{x}.
}
\description{
Constructs the expectation-maximization (EM) optimization setup.
}
