% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{get_gaussian_envelope}
\alias{get_gaussian_envelope}
\title{Construct Gaussian Envelope}
\usage{
get_gaussian_envelope(log_target_density, type = "manual", mu = 0, sd = 1)
}
\arguments{
\item{log_target_density}{Function for computing the log target density.}

\item{type}{Character string specifying envelope type, either \code{"tight"} or \code{"manual"}.}

\item{mu}{Optional mean for the Gaussian envelope.}

\item{sd}{Optional standard deviation for the Gaussian envelope.}
}
\value{
An envelope object with sampling and density properties.
}
\description{
Builds a Gaussian envelope around a target density with customizable options.
}
