% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{get_gd_setup}
\alias{get_gd_setup}
\title{Set Up GD Optimization}
\usage{
get_gd_setup(x)
}
\arguments{
\item{x}{A numeric vector of observed data.}
}
\value{
A list structure for GD containing \code{H}, \code{H_grad}, and \code{x}.
}
\description{
Constructs the gradient descent (GD) optimization setup.
}
