% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{get_log_affine_envelope}
\alias{get_log_affine_envelope}
\title{Create a Log Affine Envelope Function}
\usage{
get_log_affine_envelope(
  log_target_density,
  d_log_target_density = NULL,
  knots = NULL,
  create_sampler_function = create_sampler_function_cpp,
  create_log_envelope_density = create_log_envelope_density_cpp,
  start = 0,
  end = 1
)
}
\arguments{
\item{log_target_density}{A function that computes the log of the target density at a given point.}

\item{d_log_target_density}{An optional function that computes the derivative of the log target density. Default is \code{NULL}.}

\item{knots}{A numeric vector of knot points that define the intervals for the log affine envelope. Default is \code{NULL}.}

\item{create_sampler_function}{A function for generating samples from the envelope. Defaults to \code{create_sampler_function_cpp}.}

\item{create_log_envelope_density}{A function for computing the log envelope density for given values. Defaults to \code{create_log_envelope_density_cpp}.}

\item{start}{A numeric value indicating the start of the range for the envelope. Default is \code{0}.}

\item{end}{A numeric value indicating the end of the range for the envelope. Default is \code{1}.}
}
\value{
A list containing functions and parameters for the log affine envelope, which can be used to sample and evaluate densities.
}
\description{
This function constructs a log affine envelope for a target density, based on a specified range (\code{start}, \code{end}), knots,
and optional custom functions for creating a sampler and computing the log envelope density.
The envelope can be used for importance sampling or other density estimation methods.
}
