% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{get_log_alpha_prime}
\alias{get_log_alpha_prime}
\title{Compute Log Alpha Prime for Envelope}
\usage{
get_log_alpha_prime(log_target_density, log_envelope_density)
}
\arguments{
\item{log_target_density}{A function to compute the log target density.}

\item{log_envelope_density}{A function to compute the log envelope density.}
}
\value{
Optimal log \code{alpha} as \code{log_alpha_prime}.
}
\description{
Determines the optimal log constant \code{alpha} for comparing target and envelope densities.
}
