% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{get_optimal_parameters}
\alias{get_optimal_parameters}
\title{Optimal Parameters for Gaussian Envelope}
\usage{
get_optimal_parameters(log_target_density)
}
\arguments{
\item{log_target_density}{A function for computing the log target density.}
}
\value{
A list containing the optimal mean, standard deviation, and acceptance probability.
}
\description{
Finds the optimal mean and standard deviation parameters for the Gaussian envelope
by maximizing the acceptance probability.
}
