% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{get_proposals.naive}
\alias{get_proposals.naive}
\title{Generate Proposals Using Naive Method}
\usage{
\method{get_proposals}{naive}(setup, n)
}
\arguments{
\item{setup}{A list containing a sampler function, \code{setup$sampler}, which generates proposals.}

\item{n}{An integer specifying the number of proposals to generate.}
}
\value{
A numeric vector or list of proposals generated by the sampler function.
}
\description{
This function generates a set of proposals using a naive sampling method defined in \code{setup}.
The method leverages a sampling function (\code{setup$sampler}) that generates \code{n} proposals.
}
