% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{get_ucv_bw}
\alias{get_ucv_bw}
\title{Title: Cross-Validation Bandwidth Selection}
\usage{
get_ucv_bw(vec_obs, kernel_evaluator = kernel_cpp_epanechnikov)
}
\arguments{
\item{vec_obs}{A numeric vector of observed data points.}

\item{kernel_evaluator}{A kernel function, typically an optimized C++ version for efficiency.}
}
\value{
A numeric value representing the CV-selected bandwidth.
}
\description{
Description: Calculates the optimal bandwidth for kernel density estimation using cross-validation.
}
