% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{import_poisson_data}
\alias{import_poisson_data}
\title{Import Poisson Data and Set Data Vectors}
\usage{
import_poisson_data()
}
\value{
None. Sets global variables \code{x} and \code{z} in C++.
}
\description{
This function reads a CSV file containing Poisson data, extracts relevant vectors,
and sets them as global variables in C++ for further processing.
}
