% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{is_an_evnelope}
\alias{is_an_evnelope}
\title{Gaussian Envelope Validation}
\usage{
is_an_evnelope(
  log_target_density,
  log_envelope_density,
  lower = 0,
  upper = 10,
  n = 1000
)
}
\arguments{
\item{log_target_density}{A function to compute the log target density.}

\item{log_envelope_density}{A function to compute the log envelope density.}

\item{lower}{Lower bound of the interval to check.}

\item{upper}{Upper bound of the interval to check.}

\item{n}{Number of points to sample within the interval.}
}
\value{
\code{TRUE} if the envelope is valid; \code{FALSE} otherwise.
}
\description{
Checks if a proposed envelope function is valid, ensuring it bounds the target density.
}
