% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_target_density_cpp}
\alias{log_target_density_cpp}
\title{Log Target Density for Vector of Observations (C++)}
\usage{
log_target_density_cpp(y)
}
\arguments{
\item{y}{A numeric vector of values for which to calculate the log target density.}
}
\value{
A numeric vector of log target density values.
}
\description{
Calculates the log target density for each value in a numeric vector.
}
