% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{log_target_density_vectorize}
\alias{log_target_density_vectorize}
\title{Logarithm of Target Density Function}
\usage{
log_target_density_vectorize(y)
}
\arguments{
\item{y}{A numeric vector input for evaluating the target density.}
}
\value{
The log target density for each value of \code{y}.
}
\description{
Computes the logarithm of the target density function for Poisson-distributed data,
with vectorized operations over an input vector \code{y}.
}
