% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{log_unnormalized_gaussian_density}
\alias{log_unnormalized_gaussian_density}
\title{Log Unnormalized Gaussian Density}
\usage{
log_unnormalized_gaussian_density(x, mean, sd)
}
\arguments{
\item{x}{Numeric value to evaluate the density at.}

\item{mean}{The mean of the Gaussian distribution.}

\item{sd}{The standard deviation of the Gaussian distribution.}
}
\value{
The log of the unnormalized Gaussian density.
}
\description{
This function calculates the unnormalized logarithm of a Gaussian density given
a mean and standard deviation.
}
