% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGD.R
\name{my_optim}
\alias{my_optim}
\title{Optimization Function for Model Fitting}
\usage{
my_optim(
  X,
  y,
  basis = "non",
  p,
  lambda,
  algorithm = "batch",
  con = list(),
  function_factory = function_H_factory,
  ...
)
}
\arguments{
\item{X}{A matrix or data frame of predictor variables.}

\item{y}{A numeric vector of target values.}

\item{basis}{A character string specifying the basis function to "splines", "DR", "non".}

\item{p}{A numeric value representing parameters for the basis function if applicable.}

\item{lambda}{A numeric value specifying the regularization parameter.}

\item{algorithm}{A character string indicating the optimization algorithm "batch", "adam", "momentum", "newton", "gd".}

\item{con}{A list of control settings for the optimization process, such as lambda = 0, initial_parameters, num_samples, learning_rate, max_num_epochs, mini_batch_size, sampling_function, clipping, epsilon, patience, stopping_criteria.}

\item{...}{Additional arguments to be passed to the optimization algorithm.}
}
\value{
A list containing the optimized parameters, final loss value,
and convergence status.
}
\description{
\code{my_optim} performs optimization on input data and target values using
specified basis functions and regularization. It provides an option to
select between batch and other algorithms and allows custom control settings.
}
