% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{optimize_this.em}
\alias{optimize_this.em}
\title{EM Optimization Method}
\usage{
\method{optimize_this}{em}(setup, con, do_trace)
}
\arguments{
\item{setup}{An object of class \code{em} created by \code{get_em_setup}.}

\item{con}{A list of optimization configurations.}

\item{do_trace}{A logical, if \code{TRUE}, enables tracing during optimization.}
}
\value{
A list with optimized parameters.
}
\description{
Executes the optimization using the expectation-maximization (EM) approach.
}
