% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{optimize_this.gd}
\alias{optimize_this.gd}
\title{GD Optimization Method}
\usage{
\method{optimize_this}{gd}(setup, con, do_trace)
}
\arguments{
\item{setup}{An object of class \code{gd} created by \code{get_gd_setup}.}

\item{con}{A list of optimization configurations.}

\item{do_trace}{A logical, if \code{TRUE}, enables tracing during optimization.}
}
\value{
A list with optimized parameters.
}
\description{
Executes the optimization using the gradient descent (GD) approach.
}
