% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{optimize_this}
\alias{optimize_this}
\title{Generic Optimize Function}
\usage{
optimize_this(setup, con, do_trace, ...)
}
\arguments{
\item{setup}{An object representing the setup structure for optimization.}

\item{con}{A list of optimization configurations.}

\item{do_trace}{A logical, if \code{TRUE}, enables tracing during optimization.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Dispatches \code{optimize_this} to the appropriate method based on the object class.
}
