% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perform_em_cpp}
\alias{perform_em_cpp}
\title{Perform EM (C++)}
\usage{
perform_em_cpp(x, listOfInitialPar, maxIteration, epsilon)
}
\arguments{
\item{x}{Data.}

\item{listOfInitialPar}{List of initial parameters.}

\item{maxIteration}{Maximum iterations.}

\item{epsilon}{Stopping tolerance.}
}
\value{
A list of estimated paramters.
}
\description{
Finds parameter estimate using EM.
}
