% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGD.R
\name{perform_optim.setup}
\alias{perform_optim.setup}
\title{Perform Optimization Based on Setup Configuration}
\usage{
\method{perform_optim}{setup}(setup, algorithm, configuration, ...)
}
\arguments{
\item{setup}{list}

\item{algorithm}{algorithms.}

\item{configuration}{A list of configuration}

\item{...}{Additional arguments}
}
\value{
A list or object
}
\description{
\code{perform_optim.setup} runs an optimization routine using the provided setup
}
