% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{plot.envelope_class}
\alias{plot.envelope_class}
\title{Plot Target and Envelope Densities}
\usage{
\method{plot}{envelope_class}(object, width = 1, log_scale = T, lower = 0, upper = 0.5)
}
\arguments{
\item{object}{An object of class \code{envelope_class} containing the target and envelope density functions.}

\item{width}{A numeric value defining the additional range width for the plot beyond the peak point. Default is \code{1}.}

\item{log_scale}{A logical value indicating if the plot should use a logarithmic scale. Default is \code{TRUE}.}

\item{lower, }{upper Numeric values defining the bounds for optimization and density evaluation. Default is \code{0} and \code{0.5}.}
}
\value{
A ggplot object displaying both target and envelope densities.
}
\description{
Plots the target and envelope densities for an envelope class object, with options for logarithmic scale.
}
