% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{plot.rejection_sampler}
\alias{plot.rejection_sampler}
\title{Plot Rejection Sampling Results}
\usage{
\method{plot}{rejection_sampler}(
  object,
  density_function = NULL,
  add_normalized_density = T,
  lower = 0,
  upper = Inf
)
}
\arguments{
\item{object}{An object of class \code{rejection_sampler} containing sampling results and density functions.}

\item{density_function}{An optional function for adding a density curve. Default is \code{NULL}.}

\item{add_normalized_density}{A logical value to include a normalized target density curve. Default is \code{TRUE}.}

\item{lower, }{upper Numeric values defining the range for integrating the target density. Default is \code{0} and \code{Inf}.}
}
\value{
A ggplot object showing sampled density with optional density curves.
}
\description{
Visualizes samples from a target density, with an optional normalized density curve.
}
