% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{plot.wrapper_class}
\alias{plot.wrapper_class}
\title{Plot Combined Envelope and Sampling Results}
\usage{
\method{plot}{wrapper_class}(object, log_scale = F, add_normalized_density = T, ...)
}
\arguments{
\item{object}{An object of class \code{wrapper_class} containing envelope and sampling objects.}

\item{log_scale}{A logical value to determine if the plot should use a logarithmic scale. Default is \code{FALSE}.}

\item{add_normalized_density}{A logical value to add a normalized density curve. Default is \code{TRUE}.}

\item{...}{Additional arguments passed to the plotting functions.}
}
\value{
A ggplot object showing a combined plot of envelope and sampling results.
}
\description{
Creates a combined plot showing the envelope and rejection sampling results, using a wrapper class.
}
