% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{plot_diff.envelope_class}
\alias{plot_diff.envelope_class}
\title{Plot Difference Between Target and Envelope Densities}
\usage{
\method{plot_diff}{envelope_class}(object, width = 1, log_scale = T, lower = 0, upper = 0.5)
}
\arguments{
\item{object}{An object of class \code{envelope_class} containing the target and envelope density functions.}

\item{width}{A numeric value defining additional range width for the plot beyond the peak point. Default is \code{1}.}

\item{log_scale}{A logical value for plotting on a logarithmic scale. Default is \code{TRUE}.}

\item{lower, }{upper Numeric values for the optimization and density evaluation bounds. Default is \code{0} and \code{0.5}.}
}
\value{
A ggplot object showing the difference between target and envelope densities.
}
\description{
Plots the difference between the target and envelope densities for an envelope class object.
}
