% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{plot_target.envelope_class}
\alias{plot_target.envelope_class}
\title{Plot Target Density}
\usage{
\method{plot_target}{envelope_class}(object, width = 1, log_scale = T, lower = 0, upper = 0.5)
}
\arguments{
\item{object}{An object of class \code{envelope_class} containing the target density function.}

\item{width}{A numeric value defining the additional range width for the plot beyond the peak point. Default is \code{1}.}

\item{log_scale}{A logical value indicating if the plot should use a logarithmic scale. Default is \code{TRUE}.}

\item{lower, }{upper Numeric values defining the lower and upper bounds for the optimization range. Default is \code{0} and \code{0.5}.}
}
\value{
A ggplot object showing the target density plot.
}
\description{
Plots the target density of an envelope class object, optionally on a logarithmic scale.
}
