% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS.R
\name{rejection_sampler_wrapper}
\alias{rejection_sampler_wrapper}
\title{Rejection Sampling Wrapper}
\usage{
rejection_sampler_wrapper(setup, sample_size, type = "vectorize")
}
\arguments{
\item{setup}{Envelope object of class \code{"envelope_class"}.}

\item{sample_size}{Number of samples to draw.}

\item{type}{Sampling method, either \code{"cpp"} or \code{"vectorize"}.}
}
\value{
A list with sampled values and metadata.
}
\description{
A wrapper function for rejection sampling based on the specified envelope class.
}
