% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_data_vectors}
\alias{set_data_vectors}
\title{Set Data Vectors (C++)}
\usage{
set_data_vectors(x_in, z_in)
}
\arguments{
\item{x_in}{A numeric vector to set the global vector \code{x}.}

\item{z_in}{A numeric vector to set the global vector \code{z}.}
}
\value{
None. This function is used for its side effects.
}
\description{
Initializes the global vectors \code{x} and \code{z} with input values from R.
}
